v2/cloud_functions/requirements.txt
google.cloud
google-api-core
google-cloud-storage
google-cloud-datastore==2.0.1
google-api-python-client==2.33.0
google-cloud-pubsub
google-cloud-runtimeconfig
google-cloud-logging
pyyaml
pymysql
cryptography==3.3.2
python-jose[cryptography]==3.0.1
oauth2client
netaddr
functions-framework==3.0.0
flask==2.1.0
google-cloud-error-reporting==1.5.2